[TEST] Consolidate fragmented test files into test_typostats.py
